backend/requirements.txt
annotated-types==0.7.0
anthropic==0.93.0
anyio==4.12.1
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
cloudinary==1.44.1
distro==1.9.0
dj-database-url==3.0.1
Django==4.2.30
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docstring_parser==0.17.0
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.14.0
pillow==11.3.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
